requirements.txt
asgiref
beautifulsoup4==4.11.1
certifi
dj-database-url==1.2.0
django~=4.1
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-form==3.4
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==22.3
django-sortedm2m==3.1.1
django-widget-tweaks==1.4.12
Pillow==9.4.0
psycopg2-binary==2.9.5
pydantic==1.10.4
python-decouple==3.7
soupsieve==2.3.2.post1
sqlparse
typing_extensions==4.4.0
Unipath==1.1

django~=4.1
django-filter~=22.1